crimson: extract formatter for time_point types
